fix: verify complete follower closure across successive cuts
