+++ packages/daemon/src/gui-s3-control.ts
@@ -1,5 +1,5 @@
 import { isJsonObject, rejectSecretKeys, type JsonObject } from "./protocol/json-rpc-types.ts";
-import { validateReceiptAcceptance } from "../../kernel/src/domain/receipt-acceptance.ts";
+import { validateReceiptAcceptance } from "../../kernel/src/index.ts";
 import { isContractVersion } from "../../kernel/src/domain/contract-version.ts";
 
 export interface DaemonControlReceipt extends JsonObject {
