+++ packages/daemon/test/repo-settings.fixture.ts
@@ -3,6 +3,7 @@ import { existsSync, readFileSync } from "node:fs";
 import path from "node:path";
 import {
   compileSettingsChangedEvent,
+  daemonRegistryPaths,
   compileVerticalDeclarationEvent,
   makeTaskEventStore,
   preflightCanonicalGeneration,
@@ -10,7 +11,6 @@ import {
   registerDaemonRepo as registerProductDaemonRepo,
   resolveHarnessLayout,
 } from "../../kernel/src/index.ts";
-import { daemonRegistryPaths } from "../../kernel/src/daemon/registry.ts";
 import { defaultAssets } from "../../preset/src/preset-resolver-common.ts";
 import { canonicalRoot, workspaceId } from "../src/protocol/daemon-protocol.contract.ts";
 import { openRepoCell as openProductRepoCell } from "../src/repo-cell.ts";
